Moved the factory out of the PathBuilder and into its own class.
