+++ checker/src/main/kotlin/com/dickow/chortlin/checker/correlation/factory/PathBuilderFactory.kt
@@ -1,0 +1,10 @@
+package com.dickow.chortlin.checker.correlation.factory
+
+import com.dickow.chortlin.checker.correlation.builder.PathBuilder
+
+object PathBuilderFactory {
+    @JvmStatic
+    fun root(): PathBuilder {
+        return PathBuilder()
+    }
+}
