+++ test-application/src/test/kotlin/com/dickow/chortlin/test/application/correlation/CreateCorrelationSetsTest.kt
@@ -6,19 +6,19 @@ import com.dickow.chortlin.checker.choreography.participant.ObservableParticipan
 import com.dickow.chortlin.checker.choreography.participant.ParticipantFactory.external
 import com.dickow.chortlin.checker.choreography.participant.ParticipantFactory.participant
 import com.dickow.chortlin.checker.correlation.CorrelationValue
-import com.dickow.chortlin.checker.correlation.builder.PathBuilder.Builder.root
 import com.dickow.chortlin.checker.correlation.factory.CorrelationFactory.correlation
 import com.dickow.chortlin.checker.correlation.factory.CorrelationFactory.defineCorrelation
+import com.dickow.chortlin.checker.correlation.factory.PathBuilderFactory.root
 import com.dickow.chortlin.checker.trace.value.ObjectValue
 import com.dickow.chortlin.checker.trace.value.RootValue
 import com.dickow.chortlin.checker.trace.value.StringValue
+import com.dickow.chortlin.interception.configuration.InterceptionConfiguration
+import com.dickow.chortlin.shared.exceptions.ChoreographyRuntimeException
+import com.dickow.chortlin.shared.exceptions.InvalidChoreographyException
 import com.dickow.chortlin.test.application.instrumentation.OnlineInstrumentationTests
 import com.dickow.chortlin.test.application.shared.AuthResult
 import com.dickow.chortlin.test.application.shared.AuthenticatedService
 import com.dickow.chortlin.test.application.shared.Authentication
-import com.dickow.chortlin.interception.configuration.InterceptionConfiguration
-import com.dickow.chortlin.shared.exceptions.ChoreographyRuntimeException
-import com.dickow.chortlin.shared.exceptions.InvalidChoreographyException
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
