+++ test-application/src/test/kotlin/com/dickow/chortlin/test/application/instrumentation/ApplyInstrumentationTests.kt
@@ -4,17 +4,17 @@ import com.dickow.chortlin.checker.checker.result.CheckResult
 import com.dickow.chortlin.checker.choreography.Choreography
 import com.dickow.chortlin.checker.choreography.participant.ParticipantFactory.external
 import com.dickow.chortlin.checker.choreography.participant.ParticipantFactory.participant
-import com.dickow.chortlin.checker.correlation.builder.PathBuilder.Builder.root
 import com.dickow.chortlin.checker.correlation.factory.CorrelationFactory.correlation
 import com.dickow.chortlin.checker.correlation.factory.CorrelationFactory.defineCorrelation
+import com.dickow.chortlin.checker.correlation.factory.PathBuilderFactory.root
+import com.dickow.chortlin.checker.trace.Trace
+import com.dickow.chortlin.checker.trace.TraceElement
 import com.dickow.chortlin.interception.InterceptStrategy
 import com.dickow.chortlin.interception.configuration.InterceptionStrategy
 import com.dickow.chortlin.interception.instrumentation.ByteBuddyInstrumentation
-import com.dickow.chortlin.checker.trace.Trace
-import com.dickow.chortlin.checker.trace.TraceElement
-import com.dickow.chortlin.test.application.shared.builder.TestObservableBuilder
 import com.dickow.chortlin.interception.observation.Observation
 import com.dickow.chortlin.test.application.shared.*
+import com.dickow.chortlin.test.application.shared.builder.TestObservableBuilder
 import java.util.*
 import kotlin.test.Test
 import kotlin.test.assertEquals
